+++ src/main/java/open/commons/annotation/Setter.java
@@ -1,0 +1,57 @@
+/*
+ * Copyright 2019 Park Jun-Hong_(parkjunhong77/google/com)
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/*
+ *
+ * This file is generated under this project, "open-commons-core".
+ *
+ * Date  : 2019. 6. 20. 오후 6:06:22
+ *
+ * Author: Park_Jun_Hong_(fafanmama_at_naver_com)
+ * 
+ */
+
+package open.commons.annotation;
+
+import static java.lang.annotation.ElementType.FIELD;
+import static java.lang.annotation.ElementType.METHOD;
+import static java.lang.annotation.RetentionPolicy.RUNTIME;
+
+import java.lang.annotation.Documented;
+import java.lang.annotation.Retention;
+import java.lang.annotation.Target;
+
+@Documented
+@Retention(RUNTIME)
+@Target({ FIELD, METHOD })
+/**
+ * 데이터를 설정하는 대상을 기술.<br>
+ * 생성되는 객체의 메소드에 설정한다.
+ * 
+ * @since 2019. 6. 20.
+ * @author Park_Jun_Hong_(fafanmama_at_naver_com)
+ * 
+ * @see Getter
+ */
+public @interface Setter {
+
+    /** 데이터 이름 */
+    String name() default "";
+
+    /** 데이터 타입 */
+    Class<?> type();
+
+}