+++ src/main/java/open/commons/utils/ObjectUtils.java
@@ -26,8 +26,16 @@
 
 package open.commons.utils;
 
+import java.lang.reflect.Method;
 import java.util.Collection;
+import java.util.List;
+import java.util.Map;
+import java.util.Map.Entry;
 import java.util.function.Function;
+import java.util.stream.Collectors;
+
+import open.commons.annotation.Getter;
+import open.commons.annotation.Setter;
 
 /**
  * Object 타입의 데이터 처리를 지원하는 유틸리티 클래스.
@@ -287,5 +295,87 @@ public class ObjectUtils {
     public static <C extends Collection<Object>, R extends Collection<String>> R toString(C objects, Class<? extends R> classType)
             throws UnsupportedOperationException, InstantiationException, IllegalAccessException {
         return to(objects, classType, strFunction);
+    }
+
+    /**
+     * {@link Getter}, {@link Setter} 어노테이션이 적용된 객체를 변환하여 새로운 타입의 객체로 제공한다. <br>
+     * 
+     * <pre>
+     * [개정이력]
+     *      날짜      | 작성자   |   내용
+     * ------------------------------------------
+     * 2019. 6. 20.     박준홍         최초 작성
+     * </pre>
+     *
+     * @param src
+     *            입력 데이타
+     * @param targetType
+     *            변환 타입
+     * @return
+     *
+     * @since 2019. 6. 20.
+     * @version
+     * @author Park_Jun_Hong_(fafanmama_at_naver_com)
+     * 
+     * @see Getter
+     * @see Setter
+     */
+    public static <S, D> D transform(S src, Class<D> targetType) {
+        AssertUtils.assertNulls("'source' object or 'target' type MUST NOT be null !!!", IllegalArgumentException.class, src, targetType);
+
+        List<Method> getters = AnnotationUtils.getAnnotatedMethods(src, Getter.class);
+        if (getters.size() < 1) {
+            return null;
+        }
+
+        List<Method> setters = AnnotationUtils.getAnnotatedMethods(targetType, Setter.class);
+        if (setters.size() < 1) {
+            try {
+                return targetType.newInstance();
+            } catch (InstantiationException | IllegalAccessException e) {
+                throw new IllegalStateException(e);
+            }
+        }
+
+        // #0. Setter 메소드 재정렬
+        final Map<String, Method> setterMap = CollectionUtils.toMapHSV(setters, m -> {
+            Setter annoSetter = m.getAnnotation(Setter.class);
+            return String.join("-", annoSetter.name(), annoSetter.type().toString());
+        }, m -> m);
+
+        final Function<Method, String> GetterKeyGen = m -> {
+            Getter annoGetter = m.getAnnotation(Getter.class);
+            return String.join("-", annoGetter.name(), annoGetter.type().toString());
+        };
+
+        // #1. Setter와 동일한 식별정보를 갖는 Getter 추출
+        List<Method> gettersFiltered = getters.stream().filter(m -> {
+            return setterMap.containsKey(GetterKeyGen.apply(m));
+        }).collect(Collectors.toList());
+
+        // #2. Setter와 연결되는 Getter 메소드 재정렬
+        Map<String, Method> getterMap = CollectionUtils.toMapHSV(gettersFiltered //
+                , m -> {
+                    Getter annoGetter = m.getAnnotation(Getter.class);
+                    return String.join("-", annoGetter.name(), annoGetter.type().toString());
+                }, m -> m);
+
+        // #3. 객체 생성
+        D target = null;
+        try {
+            target = targetType.newInstance();
+        } catch (InstantiationException | IllegalAccessException e) {
+            throw new IllegalStateException(e);
+        }
+
+        for (Entry<String, Method> entry : setterMap.entrySet()) {
+            try {
+                entry.getValue().invoke(target, getterMap.get(entry.getKey()).invoke(src));
+            } catch (Exception e) {
+                throw new IllegalArgumentException(e);
+            }
+        }
+
+        return target;
     }
 }